BUG#34538796 ANSI_QUOTES for sql_mode breaks \status command in MySQL Shell
Bug was caused by incorrect written sql character set query, where a space was added between version strings with double quotes ("). When ANSI_QUOTES is enabled, double quotes in selection query is treated by the MySQL server as an identifier for columns, and the same error is reported as using backtick quotes (`). The bug was corrected on MySQL Shell side, by replacing with single quotes (') that are not used in identifiers. Added tests for \status command with ANSI_QUOTES enabled and disabled.

Change-Id: I1ff9c9f121d90a78eb0d642ba0f05023f24f17c5