BUG#34747533 el7 rpm needs openssl 1.0.2 but requires 1.0.any
The rpm packages were not requiring a 1.0.2 OpenSSL, but instead allowed
to be installed even if only 1.0.1 was present in the system.

CMake scripts have been adjusted in order to ensure that bundled shared
libraries are installed with executable permissions.

Additionally, fixed dependencies of bundled binaries.

If we bundle a binary with Shell, and at the same time we bundle
OpenSSL, then the bundled binary should use it instead of the system
one.

Change-Id: Ib53ff04a56fba0e3534e89c4db995d60b3041b4d