BUG#34709673 can't USE schema with " in name while ANSI_QUOTES is on
In SQL shell mode "use" command is treated like a shell command, and
parameter extracting is handled by split_command_line method.
This method was not aware of ANSI_QUOTES mode, and did not have
functionality to escape double double-quote characters when this mode
was enabled. Definitions of "use" command was changed, to avoid parsing
of parameters by split_command_line on SQL shell mode. Added tests, to
verify change.

Change-Id: I05046693b8a918bbf97119c595e0183819c083e5