BUG#34647972  Shell sets super_read_only on multi primary cluster member
One of the requirements of WL#11561 was to make sure that the
super_read_only was properly set after rejoining a member to a given
cluster. This was implemented without looking at the cluster topology,
meaning that, in a multi-primary scenario, all instances (except one)
would end up with super_read_only enabled.

This patch fixes this issue by not changing the super_read_only variable
of an instance if the cluster has a multi-primary topology.

Change-Id: Ib600ffbfac33e59442a982ade8a1b9d24abe6001