BUG#34530914 ReplicaSet.status: Failed to execute query on Metadata Illegal mix of collations
When the server is configured with "character-set-client-handshake=OFF"
(or --skip-character-set-client-handshake), this means that the server
collation is used in the client's sessions (instead of being
negotiated). This can cause problems in some queries if they aren't
prepared to deal with a different collation.

This patch fixes this scenario by explicitly setting the collation in
the sessions used by the AdminAPI.

Change-Id: I3da203b3851e7ef3aa377a6a14f2e3481b8a5f31