Remove extra error message
In WL11969, an additional error message was added when trying to acquire
the Primary of a ReplicaSet. Since the tests usually don't check for
erros in the shell ouput, just if the command succeeded or nor, this
went by unnoticed until QA reported it.

Simply removed the unnecessary error message.

Change-Id: Ic1d4c61eee03c9892cfefa45508301aa1b560a65