BUG#34768224 loadDump fails on view that references a view whose definer does not yet exist
Loading a dump which contained a view which referenced another view
which had a DEFINER clause set failed, because user used as the definer
did not exist. As a fix, users are now loaded after all objects and view
placeholders are created, but before the view placeholders are replaced
with actual views.

Change-Id: Iac698a1081030e6bd9a20647be289290c5b76727