Fix random segfault when running tests on EL7 platform
Random segfault when running the
Shell_scripted/Auto_script_py.run_and_check/column_metadata_all_types_norecord
test has been fixed. Crash was caused by the Types_cpp tests, which
cleared the whole metadata table. If a Type constant existed before this
happened, it would store a pointer to the metadata of 'help' function,
which was no longer valid. Accessing this constant could lead to a
crash, depending on the current contents of the memory.

Change-Id: I8f07db9187ccd2075a5a42267a79dbb03754cd5a