Fixing shell crash caused by polling error on SSH connections.
When a polling error occurs at Ssh_tunnel_handler::handle_connection
the a call to cleanup_event(...) will be done.

If the session is either not connected and fails to reconnect the ssh
tunnel thread will be terminated.

If at that point the shell session is closed the destructor will trigger
a new call to cleanup_event(...) which will cause the shell to crash.

Change-Id: I8b1293036dc8ec1fcbb2bba40d8186241dabeb21