BUG#34787542 Error creating a MySQL 8.0.22 cluster with the latest InnoDB Cluster version
The patch for BUG#31798724 and BUG#31467823 ensures that the new
terminology is used on replication commands according to the target
Server version. However, the patch did a wrong comparison of the version
and considered that the new terminology is already available since
8.0.22 when, in fact, it's only 8.0.23.

That results in replication errors while using the AdminAPI on a Shell
version >= 8.0.22 managing MySQL Servers running 8.0.22.

This patch ensures the new terminology is only used if the target
Servers version is >= 8.0.23.

Change-Id: I8f047346423ae0c647e8ba73033c08575e89d37d