Bug#34787737 dissolve ClusterSet members
Bug#34787737	Dba.dropMetadataSchema: This function is not available
through a session to an InnoDB Cluster that belongs to an InnoDB
ClusterSet (MYSQLSH 51601)

dissolve and dropMetadata are now allowed on a cluster that belongs to a
ClusterSet if they're the last/only cluster or if they are invalidated.

dissolve was also performing a sync of each member with the PRIMARY,
but that is unnecessary and a simple wait for received GTIDs to be
applied is enough.

Change-Id: Ieb5d6dd8a19fb7a6d37e24c0f919556ed07ab035