AdminAPI code refactor:
The AdminAPI commands have been implemented using different
approaches/techniques, being the most used one the implementation of
Classes deriving from Command_interface. The interface was supposed
to generalize and standardize the behavior of each individual command
and workflow.

However, this has been proven to be over-engineered and not
future-proof, bringing unnecessary complexity and limitations to the API
implementation. On top of that, multiple commands share common methods
and attributes which could be grouped together but using the
Command_interface they can not.

To overcome such limitations and simplify the general approach to the
AdminAPI implementation, a new design approach is introduced by this
patch. The approach is a compile-time variant of the strategy pattern,
called policy-based design.

It allows encompassing commands that share a common behavior/feature
under the same "host" class that takes an input parameter when
instantiated that will then implement the desired interface. The input
parameter is the type of class to be used, for example
cluster::Add_instance.

The host class specifies how the main operation (run()) is meant to
behave as well as the rollback mechanisms but then is up to each
"policy" class to implement it the way it needs. Each "policy" class is
obliged to implement do_run().

In general, this allows modularization while allowing the removal of
unnecessary interfaces for a much cleaner code.

This patch sets the base for the new approach by refactoring the
following commands:

  - dba.rebootClusterFromCompleteOutage()
  - <Cluster>.addInstance()
  - <Cluster>.removeInstance()
  - <Cluster>.rejoinInstance()

Under the host class named Cluster_topology_executor. Leaving the door
open for future clean-ups and the refactoring of commands'
implementations.

Change-Id: Ic1603d55019a9e3e4bfecd7c286ac9ae5aa100be