Bug#34692990	ClusterSet primary purged binlog, rebootClusterFromCompleteOutage() wait forever
Rebooting a cluster with >1 member to a clusterset while the rebooted
cluster is missing purged transactions would freeze during sync.
The freeze happened because all members of the cluster were being synced
with the global primary and there was no check for purged transactions
that would never get replicated.

Issue was solved by checking for purged transactions before cluster
rejoin. Additionally, a replica cluster sync is done in 2 steps: the
PRIMARY of the cluster is syncted with the clusterset primary first,
then the rest of the group waits for the transactions to be applied.

If there are both errant and unrecoverable transactions, then an
errant transaction error is thrown, because fixing the errant
transaction issue through clone implies a potential data loss.

Change-Id: Ib84897fb6def93043e8a54f5dfb2f8f9a1587e29