BUG#34684193 Check for identifiers starting with a dollar sign in upgrade checker
WL#15422 deprecates unquoted identifiers starting with a dollar sign.
This commit updates the update_check utility to find schemas, tables, views,
columns and routines and present to the user a list of objects that have
an identifier starting with a dollar sign but not ending with it.
Routines are additionally checked in their definitions.
Modifications add sql quaries in new uprade_check and tests for them.

Change-Id: Ic6f8972636fd2d5500bbe2b67ce25f7c3852fb45