BUG#34433132 MySQL Shell does not prevent about orphaned routines
When there is an orphaned routine (routines which schema were somehow
dropped), it will make the MySQL server upgrade procedure fail. This commit
adds a check to upgrade_check utility, that looks for routines which schema
is not present in current instance of MySQL Server and reports them as Errors
that should be addressed.

Change-Id: If26d29944d4d24edf608a9e32f4d12f29c92381d