BUG#34764157 dumpInstance retains but comments out some grants [noclose]
The dumpInstance() utility was commenting out grant statements
referring to existing views or routines.

The dumpInstance() should only comment out grants which refer to objects
not included in the dump. With this fix, grants are commented out as
expected, though grants for missing objects are still commented out.
This is going to be fixed in another commit.

Change-Id: I250f79f8d39e9f2c8100679154960249a9d909c1