Bug#34704392 When SAS token configuration is invalid a better error needs to be made
Dump and Load operations in Azure support using two different types of
SAS Tokens: account and container SAS Tokens. Such tokens have minimal
requirements regarding the allowed services, resource types and
permissions, if the SAS token was configured incorrectly on creation,
the Dump and Load operations may fail with errors that are not
descriptive about the real root cause.

This patch introduces specific error messages for the minimal
requirements on each of the supported SAS Tokens.

Change-Id: I39a6ee07cba4d0c6d32f438bb3001d2d85b88ea8