Bug#34379393 Python Shell API Row - extra underscores in property names querying 8,0 servers
The problem is caused because in 8.0 the column names are in uppercase,
the python naming enforcement logic was converting the names to
snake_case adding an underscore before each uppercase letter.

This patch fixes the problem by exposing the fields as properties
using the exact column name for both JavaScript and Python, this is,
the column name takes precedence over the naming conventions.

Change-Id: I5b447a68a4d5f427c369dbca1cc5c56b50cd4483