BUG#34680289 no information about hidden cluster members in status()
<Cluster>.status() does not include any information about members which
have been hidden from Router traffic using the tag '_hidden'. This
information is only available via .options().

Considering that this setting is quite impacting, it should be included
in the output of .status() and .describe commands. For that reason, this
patch ensures a new attribute named 'hiddenFromRouter' is displayed in:

  - <Cluster>.status()
  - <Cluster>.describe()
  - <ReplicaSet>.status()

... for every instance that has the tag '_hidden' enabled.

Change-Id: I7e459f65d3d165d501c444fc1da19729574eb2a6