Bug#34693099	ClusterSet metadata is not rolled back when removeCluster() failed.
Removing a cluster from a ClusterSet while that cluster is down but
GR is still trying to go ONLINE would fail while trying to reset
auto-failover settings, despite the force flag being enabled. Moreover,
the rollback process would also fail because it would assume GR is running
in multiple places. And when succeeded, the metadata would not be left
in the exact same original state.

removeCluster() was fixed to handle the case of the GR plugin being
in the middle of starting up by stopping GR. Errors will also now
be ignored if force is enabled.

The rollback process for removeCluster() was changed in order to
not depend on GR being ONLINE and instead, take snapshots of
server data being modified and re-applying them when needed.

New SQL undo infrastructure was added to support snapshotting and
re-applying server changes such as delete from table, update table
and drop user.

Change-Id: Ie3348fc9a7d61d506279d2c153225e7a7b3d4354