BUG#34858763 Allow to use non-default location of the OCI config file with authentication_oci
Previously, when using the authentication_oci_client authentication
method to connect to a MySQL instance, the OCI configuration file had
to be present in the default location.

With this fix, the OCI configuration file specified using the
'oci.configFile' MySQL Shell option is used instead. If this option is
set to an empty string, the default location is used.

Change-Id: Iebbd1d21b2a74b6432d805c92bd886c5c0b41e36