Bug#34837601	CREATE REPLICA CLUSTER() FAILURE WITH MALFORMED GTID SET
createReplicaCluster() was failing with a "malformed GTID SET" error
coming from MySQL. This was triggered by sql_mode being set to
NO_BACKSLASH_ESCAPES. Some of the connections that are reopened after
a clone were not being properly initialized with a clean sql_mode.

Reconnect code was cleaned up and moved to a common place.

Change-Id: Ie50a8de25f898fbddf94e8e03312783cc8ea4689