BUG#34876423 Can't restore loadDump with deferTableIndexes=all by MySQL Error 1061
Loading a dump with the 'deferTableIndexes' option set to 'all', if it
contained a table with multiple indexes, including one which was
specified for an AUTO_INCREMENT column, failed when loader was
attempting to recreate an index.

When loader is extracting indexes from the CREATE TABLE statement it
merges two index definition clauses, if they are specified one after
another. Loader also does not extract indexes specified on an
AUTO_INCREMENT column, trying to avoid creating an invalid SQL
statement. If table had two indexes specified one after another, with
the second one being an index on an AUTO_INCREMENT column, the two
definitions were merged into one, then the definition was ignored, as
the merged fragment contained an AUTO_INCREMENT column.

As a fix, the order of operations has been reversed: first, the index
definition is checked for an AUTO_INCREMENT column, then the definitions
are merged.

Change-Id: I6803347508ea62926a0c240920774e35d8cbf90f