BUG#34840953 support external credential provides for load/dump
Added support to dump/load/export/import utilities for the
'credential_process' setting stored in the AWS `config` file.

The process is expected to write a JSON object to the standard output in
the following form:
{
  "Version": 1,
  "AccessKeyId": "AWS access key",
  "SecretAccessKey": "secret key associated with the AWS access key",
  "SessionToken": "temporary AWS session token, optional",
  "Expiration": "RFC3339 timestamp, optional"
}

The `Expiration` key, if given, specifies when the credentials are going
to expire, they will be automatically refreshed before this happens.

This credential provider has lower priority than the `credentials` file,
but higher than the `config` file.

For more information please consult the AWS docs:
https://docs.aws.amazon.com/sdkref/latest/guide/feature-process-credentials.html

Change-Id: Ia19cfbc85d935c69dba7ec2c3d59044820a6996a