BUG#34711038 InnoDB multi-primary auto_increment_increment, auto_increment_offset bug
When adding a member to a Cluster, the logic that calculated the number
of elements in the Cluster (which would then be used to decide on the
correct values of the two auto_increment_ variables) was off by 1. This
meant that, when adding the 8th member to the Cluster, the new member
would be configured with the correct values, while the already existing
members would keep using the configurations for a 7 member Cluster.

Also corrected the scenario in Cluster.rejoinInstance() when the number
of members passes 7.

Change-Id: I1d89bcdd224a567194715888c1ba7bd32828fee5