BUG#34569971 dba.checkInstanceConfiguration report: required value = configured, yet error
Tweaked how the log_bin related variables are printed in
dba.checkInstanceConfiguration() so that they show more meaningful
strings taking into account the types of variables that they are and
their context, mainly:
 - log_bin shows "<not present>" or "OFF" (for current value) and "ON"
for required value.
 - disable_log_bin and skip_log_bin show "<present>" and "<remove>" (for
current and required values) and the note was updated to "Update and
restart the server".

Change-Id: Ide0ba368304452360bee58d234d51beb697365a6