BUG#31357039 Missing or wrong IP whitelist not handled in AdminAPI
When the IP 'ipAllowlist' option is not set, GR automatically creates
one by scanning the active interfaces. The resulting list usually
includes the RFC defined private IPv4 and IPv6 networks as well as the
localhost.

If the target instance local address (generated by the AdminAPI or
specified using the option 'localAddress') isn't compatible with the
previously mentioned list, the instance won't bootstrap or join the
group.

This patch adds a check to dba.createCluster() and cluster.addInstance()
to verify that the local address (supplied or generated) is compatible
with the automaticly generated list of allowed IPs (if the list is
supplied manually, the test is skipped). Also, this only applies if the
comm stack is "XCOM", since "MYSQL" ignores the ipAllowlist.

Change-Id: Ief6d03d3a20c43caed1bc2a2a1b65bc7899dc1db