BUG#34979347 Misleading help on plugin functions with same name in different objects
The help data is registered using keys in the format of:

<object>_<member>_<suffix>

This works ok for the shell itself as the combination of those items is
unique among all the exposed APIs, however, when it comes to plugin
implementation the possibility that such uniqueness is honored gets
reduced and in case of duplication, the help for the last registered
item overrides previously registered help data, causing the shell to
produce misleading help.

This patch avoids the problem by changing the way the help data is
registered for plugins. It replaces the <object>_<member> prefix
by the fully qualified id of the member which is guaranteed to be
unique. The help lookup logic was also updated to first attempt to
resolve the help using the fully qualified id of the target item and
falling back to the old logic in case the help was not found.

Change-Id: I68b61a2588fa1b926489886bb742886dc2dc9f8e