BUG#34566034 util.loadDump errors on index creation with existing objects
When loading a dump using options:
  "deferTableIndexes": "all", "ignoreExistingObjects": true
into an instance which already contained tables which were being loaded,
operation failed when loader was trying to re-add existing indexes.

As a fix, if the `ignoreExistingObjects` option is set to true, existing
tables are checked for indexes, and only the ones missing are re-added.

Change-Id: I3ae70439a41bc236dde652c13c8c5211d7433d1b