BUG#34535236 rejoinCluster on an offline cluster needs better guidance
When trying to rejoin a Cluster back into a ClusterSet that is currently
OFFLINE (all members have GR stopped but are reachable), the error
message shown implies that the primary isn't reachable which is
incorrect. The Cluster is reachable however, it suffered a complete
outage.

This patch improves the error message by explicitly stating that the
error is due to the state of the cluster and suggests how to proceed
from there.

Change-Id: I18653b70ecfa85c258e35a0d869afa90eccc7755