Bug#35020175 -u/--user and --password options are ignored when specified before URI
Shell option handling overrides connection options with
options that appear after them. If the option that appears later is a
URI, it was being filled in with defaults, including user and
password. This made it not possible to specify user/pass before a URI
without these (e.g. -uroot -ppass mysql://localhost).

cmdline URI handling was changed to NOT fill-in defaults, since any missing
options can be filled in after all user options are procesed.

Also, password/mfa_password handling in Connection_options was changed
from treating mfa_password[0] as the same as simple password at the
getter to doing the resolution at the setters, ensuring that stored data
is consistent at the source.

Change-Id: I71b0ff9f2a9dfedf777d1860f3a0f2b0c6675545