BUG#35046654 "Incorrect recovery account" after upgrade to 8.0.32
In BUG#33235502, a new warning message was added to status() to check if
the recovery account in use for a given instance was the correct one. If
it wasn't, then rescan() would fix this error. However, the new check
didn't take into account the old recovery account prefix, while the
rescan() does. In other words, status() complains about the mismatch but
then rescan() does nothing about it.

There's no underlying issue with the account, the only thing that needs
to be fixed is the message itself. This patch does this by updating the
check so that it recognizes old / legacy recovery account formats.

Change-Id: I110d159fd6190cf58ca85f115d6282219a43737a