Fix connectivity checks
In WL#13688, some connectivity checks were introduced to test the new
SSL configurations. These tests use the localAddress, but were ignoring
the port, which lead to problems when the localAddress used any port
other than the default one.

This patch fixes the connectivity checks by parsing the address and
correctly using the port specified in it.

Change-Id: Icfd588cfc8b06aaaf1f8cd70a71137840f721a75