BUG#34952027 util.dumpInstance comments out grants on wildcard schema names
When dumping user grants which contained privileges at the database
level on schema names with wildcards, GRANT statements were commented
out in the dump.

As a fix for all GRANT-related issues, filtering of the grant statements
based on the include/exclude options has been removed entirely from both
dumper and loader. Dumper will print a warning if a grant statement on
an excluded object is detected.

Additional changes:

When dumping with the `ocimds` option set to true, all database level
grants containing wildcard characters `_` or `%` are reported as errors,
because the MDS instances have the `partial_revokes` system variable set
to ON. Setting this variable causes the wildcard characters to be
treated literally, which may be unexpected. These issues need to be
fixed manually (see https://dev.mysql.com/doc/refman/en/grant.html),
then dump should be rerun with a new "compatibility" option
"ignore_wildcard_grants", which silences the errors. Alternatively,
problematic accounts can be excluded from the dump using the
"excludeUsers" option, or dumping the users can be disabled entirely.

A new option "handleGrantErrors" option accepting values: "abort",
"drop_account" and "ignore" has been added to the util.loadDump(). It
controls the action taken in case of errors related to the GRANT/REVOKE
statements:
 - "abort": throws an error and aborts the load
 - "drop_account": deletes the problematic account and continues
 - "ignore": ignores the error and continues loading the account
The default value is: "abort".

The documentation of the "ignoreExistingObjects" load option has been
improved, added an explanation that the loader checks for existing user
accounts and DDL objects and fails if any existing account or object is
found. Setting this option to true suppresses this error (duplicate
objects are still reported), but CREATE statements are executed anyway.

Internal class which handles user privileges has been updated to handle
database-level wildcard grants.

Change-Id: I0c66b066e7b7657432d2c49fcc1e04f8a6177dec