Fixed calling non performance_schema tables on testutil.waitMemberState
The latest commit added code in testutil.waitMemberState to check if the
member was part of a replicaset or not. To do this, however, we need to
read from the MD, which is sometimes not possible (e.g.: because of
group_replication_consistency).

This patch fixes that by putting that logic in a new command to check
the state of a given replication channel.

Change-Id: I0ea8757fa0acc2c4f7145839bf516a8d5efc2b95