BUG#31711835 dba.upgrademetadata() fails with "error: truncated incorrect integer value: 'true'" when cluster was created with shell 1.0.x
When upgrading to metadata 2.0, the value of the attribute "adopted" (in
the legacy mysql_innodb_cluster_metadata_previous.replicasets table),
which was stored as a string, is converted to an unsigned integer. This
conversion, however, assumes that the string only has a numeric value,
like "1" or "0", and doesn't consider "true" or "false". This leads to
the error reported.

This patch updates the upgrade script, so that "true" and "false" are
now also parsed and converted correctly.

Change-Id: I5ee3ee03118a762f619f170e1958b36291ae0545