Refactored cluster/replicaset scripting objects
Created a common base class for Cluster, ClusterSet and ReplicaSet
for the scripting interface and moved identical functions to it:
        setupAdminAccount(), setupRouterAccount(), listRouters(),
	setRoutingOption(), routingOptions()

Note: ClusterSet::listRouters() has a slightly different interface,
although it probably shouldn't.

Also added support for generalized help text for common functions,
allowing for keywords like the topology type to be abstracted.

Change-Id: I854c01fbb48e189b203c5557f816c8696ca8f6cc