BUG#34778797 offline-mode is persisted when doing dba.rebootClusterFromCompleteOutage
When adding or rejoining instances to a Cluster (and by extension when
rebooting one), shell checks the value of "offline_mode" and, if
enabled, disables and persists the variable. This overwrites the value
of the variable, which the user might not be expecting if it added
the variable in my.cnf explicitly.

This patch changes this behaviour by only changing the value of the
variable GLOBALly (and not persist it). Also, a new warning was added to
status checking if the variable is persisted and enabled on an instance,
so that the user is aware of the possible risks of doing so.

Change-Id: If319fe6c85c9b17e957ff610c39179f2b1e4134b