BUG#35046432 Segmentation fault when calling add_instance on a disconnected cluster
In Javascript mode, the shell, before any Cluster command is executed,
checks if the current session is valid (i.e.: the cluster wasn't
disconnected). This check, however, isn't performed in Python mode for
every command, which means that, if the command (like
Cluster.add_instance) is called on a disconnected cluster, then tries to
use an invalid session, causing the crash.

This patch fixes this by making sure that all commands in Python check
if the cluster is connected before proceeding.

Change-Id: I0c8add8582617eca39be4002ad832aa8bc7c4ad2