WL#15571 BUG#35115658 Query attributes support in sessions through the MySQL protocol
This patch introduces two methods to define query attributes in the
MySQL Shell:

- \query_attributes shell command
- ClassicSession.setQueryAttributes

The new shell command receives a list of space separated attribute/value
pairs while the API function receives a dictionary with the attribute
definitions.

The attributes will be defined on the next SQL executed by the user.

Change-Id: Iba42e1acbe793430f13e57d1981b21210c8fbd66