BUG#34614769 clusterset.status() reports repl channel issues when its own failover is ongoing
Currently, shell retrieves the state of the replication channel
"clusterset_replication", and if it contains an error, it reports it as
such. This means that, if the channel is connecting (and the error is
related to that), it will never show the state as connecting but always
as in error. This is misleading because errors during a connection
attempt are acceptable.

This patch addresses this by acknowledging that the channel is
connecting and ignoring the error until the channel state updates to
either ON or OFF. This means that, in ClusterSet.status(), the
"clusterSetReplicationStatus" field in a cluster can return "CONNECTING"
(in which case the "globalStatus" is now "OK"), and in
ReplicaSet.status(), the "status" field of an instance can now show
"CONNECTING".

Change-Id: Ief5d96f14ce6342a29f78916f316f99c9188116d