BUG#33930873 Set group_replication_paxos_single_leader using MySQL Shell
In MySQL 8.0.27, Group Replication introduced the ability to configure
the Group to use a single leader when driving consensus in Paxos, if
in single-primary mode. Using the option is not straightforward in
several scenarios and the AdminAPI was lacking support for it.

This patch addresses that by adding support for enabling/disabling Paxos
Single Leader in InnoDB Cluster and providing automatic management of it
in the Cluster/ClusterSet operations.

The support is only available when using MySQL Server version 8.0.31 or
higher, because the AdminAPI needs the information provided by the column
`WRITE_CONSENSUS_SINGLE_LEADER_CAPABLE` in
performance_schema.replication_group_communication_information,
introduced in 8.0.31.

In general, this patch:

  - Adds support for enabling/disabling Paxos Single Leader in
  dba.createCluster() and ClusterSet.createReplicaCluster()
  - Adds support for changing the setting while recovering a Cluster in
  dba.rebootClusterFromCompleteOutage()
  - Generates and sets the right value for the setting automatically
  when adding or rejoining members to Clusters
  - Lists the value of the setting in use in Cluster.options() and
  Cluster.status({extended: 1}) and ClusterSet.status({extended:2})

Change-Id: I55e61c0a1a3b6a586d29570f32dfe3ecb1a8b8d2