BUG#35000998 CLUSTERSET.SETPRIMARYCLUSTER FAILING DUE TO ERRANT TRANASCTION
A mismatch of the value stored in the Metadata for
group_replication_view_change_uuid and the current runtime value results
in errors during switchovers and failovers in ClusterSet since the
reconciliation of the transaction sets regarding VCLEs cannot be done if
the current UUID does not match what's been stored in the metadata.

This patch fixes that by adding a new validation in Cluster.status() to
confirm the runtime value of the sysvar matches the one stored in the
metadata and if not add a warning to let users know about it and
indicate the usage of Cluster.rescan() to fix it.

Cluster.rescan() is also extended to detect this scenario and update the
metadata accordingly.

Change-Id: I0e1ed9aaa4c00da0bcc567037311c86d0d2c0eda