BUG#35140916 Improve Shell Warning For Bootstrapped Routers before ClusterSet creation
When upgrading a standalone InnoDB Cluster to a ClusterSet any Router
instance bootstrapped against it will keep working, however, it won't
recognize the ClusterSet and won't have the optimal configurations in
place to work on ClusterSet topologies.

The AdminAPI warned about the configurations but didn't warn about the
Router not recognizing the ClusterSet and acting as if the Cluster
remained standalone.

Users need to be warned about it, otherwise, there's the risk of missing
that information and assuming Routers will automatically recognize the
ClusterSet when it won't and unexpected behaviors will be reported
during ClusterSet operations.

This patch fixes that by extending the warning messages to include the
missing information.

Change-Id: Ie5068d244b72ceeb0e2d98079ec81b1237e612be