Fixed unintended behaviour change in connection
Default connection when no host or port or socket is specified used to
be a X protocol session via TCP/IP, but that unintentionally changed to
X protocol via socket.

This patch ensures that by default, TCP/IP connections are used for X
protocol (but not classic) to avoid issues in systems where the default
socket path does not match the one compiled in.

Change-Id: Ia7a5f34f7092f1085d521131c40abee820049b6c