WL#15056 AdminAPI: InnoDB Cluster Read Replicas
Support for adding Read Replicas in InnoDB Clusters, using the
AdminAPI.

Read Replicas are the means to scale out read traffic while offloading
other Cluster members and additionally providing more data redundancy
to the dataset.

New commands
============

<Cluster>.addReplicaInstance(instance[, options])
-------------------------------------------------
Adds a Read Replica Instance to the Cluster.

Parameters:

- instance: host:port of the target instance to be added as a Read
Replica.
- options: Dictionary with additional parameters described below.

Options:

- dryRun: boolean if true, all validations and steps for creating a
Read Replica Instance are executed, but no changes are actually made.
- label: an identifier for the Read Replica Instance being added, used
in the output of status() and describe().
- replicationSources: The list of sources for the Read Replica
Instance.
By default, the list is automatically managed by Group Replication and
the primary member is used as source.
- recoveryMethod: Preferred method for state recovery/provisioning. May
be auto, clone or incremental. Default is auto.
- recoveryProgress: Integer value to indicate the recovery process
verbosity level.
- timeout: maximum number of seconds to wait for the instance to sync
up with the PRIMARY after it's provisioned and the replication channel
is established. If reached, the operation is rolled-back. Default is 0
(no timeout).
- cloneDonor: The Cluster member to be used as donor when performing
clone-based recovery.

<Cluster>.setRoutingOption()
----------------------------
Changes the value of either a global Cluster Routing option or of a
single Router instance

Parameters:

- router: Identifier of the target router instance
(e.g. 192.168.45.70::system).
- option: The Router option to be changed.
- value: The value that the option shall get (or null to unset).

Options:

- tags: Associates an arbitrary JSON object with custom key/value pairs
with the Cluster metadata.
- read_only_targets: Routing policy to define Router's usage of Read
Only instance. Default is 'secondaries'.

<Cluster>.routingOptions()
--------------------------
Lists the Cluster Routers configuration options.

Parameters:

- router: Identifier of the router instance to query for the options.

Extensions
==========

<Cluster>.removeInstance(instance[, options])
---------------------------------------------
Extended to support Read-Replicas too.

New options:

- dryRun: to perform all validations and log the changes but not
execute them.
- timeout: maximum number of seconds to wait for the instance to sync
up with the PRIMARY. If reached, the operation is rolled-back. Default
is unlimited (zero).

<Cluster>.rejoinInstance(instance[, options])
---------------------------------------------
Extended to support Read-Replicas too.

New options:

- dryRun: to perform all validations and log the changes but not
execute them.
- recoveryMethod: to select the provisioning method used: incremental,
clone or auto. The option behavior is equal to Cluster's or
ReplicaSet's behavior but the default value is incremental.
- recoveryProgress: Integer value to indicate the verbosity level of
the recovery process of the Read Replica being provisioned.
- cloneDonor: The Cluster member to be used as donor when performing
clone-based recovery (available only for Read Replicas).
- timeout: maximum number of seconds to wait for the instance to sync
up with the PRIMARY after it's provisioned and the replication channel
is established. Default is unlimited (zero) (available only for Read
Replicas).

<Cluster>.describe()
--------------------
Extended to include Read-Replicas in the topology description output.

<Cluster>.status()
------------------
Extended to include Read-Replicas in the topology status output.

<Cluster>.dissolve() and <ClusterSet>.removeCluster()
-----------------------------------------------------
Also remove Read-Replicas from the Clusters.

<ClusterSet>.setRoutingOption()
-------------------------------
Extended witha new option: 'read_only_targets'

ClusterSet>.routingOptions()
----------------------------
Extended to include the option 'read_only_targets' in the output.

dba.createCluster() and <Cluster>.createClusterSet()
----------------------------------------------------
Set read_only_targets to secondaries as the default value in the
Metadata.

<Cluster>.forceQuorumUsingPartitionOf()
---------------------------------------
Extended to stop replication on all reachable ONLINE Read-Replicas of
the Cluster and when quorum is restored, restart it.

<Cluster>.setInstanceOption()
-----------------------------
Extended with a new option 'readReplicaReplicationSources' to allow
changing the replication sources configured for a Read-Replica

dba.getCluster(), dba.getClusterSet(), <Cluster>.getClusterSet()
----------------------------------------------------------------
Extended to allow obtaining the topology from a Read-Replica.

<Cluster>.fenceAllTraffic()
---------------------------
Extended to enable offline_mode on all Read-Replicas of the Cluster.

dba.rebootClusterFromCompleteOutage()
-------------------------------------
Extended to not automatically rejoin Read-Replicas to the Cluster.

Other changes
=============
Metadata version bumped to 2.2.0, in order to support Read-Replicas.

dba.upgradeMetadataSchema() extended to support 2.2.0.

Change-Id: Iaf6b56cd48054b5edd36e0ba1a1a46ccbaab25aa