Fix test failure in macos
Prevent scoped locks and callbacks from being invoked prematurely.

The defaulted move operators and ctors apply std::move for each class
member, which apparently, in MacOS build with clang, which leaves
std::function in an undefined state that depends on the platform
(unchanged in macos clang, null in gcc). This means that its
"operator bool()" can't be used to check if it's valid or not.

In MacOS, this causes the callback to be triggered prematurely.

Change-Id: I208e34726f97904e8233cdc6fa3ebe95ec687dc6