Inline variables declared in header files
Variables declared with "static", "const" (or "constexpr") in header
files create copies of said variable in every compilation unit the
header ends up being used. While for some types of variables this
duplication is negligible, for containers with lots of entries this can
accumulate.

In this instance, inlining std::set declared in an header files saves
nearly 2MB of RAM (on a total of 4MB) on static initialization (before
main() is even called).

This patch also cleans up some header declarations.

Change-Id: I7f148328b8eb1acc4f29e91fbb21567809a5eab7