BUG#35313366 Shell crashes when import_table() imports single uncompressed file
When importing a single uncompressed file using util.import_table(), if
reading the imported file resulted in an exception, Shell would crash.

The reader thread which is used in this scenario did not handle
exceptions, an exception which happened during read was not caught and
caused a crash.

This thread is now protected with try/catch, any exceptions are properly
reported to the main thread.

Change-Id: Idfa7b44558ffd42541ab61c0b234badca0f15e4b